--- front/src/components/NavbarComponent/Navbar.js
@@ -1,18 +1,0 @@
-export default {
-  name: 'NavbarComponent',
-
-  props: ['activeMenu'],
-
-  data() {
-    return {
-      isOpen: false,
-    };
-  },
-
-  methods: {
-    logout() {
-      this.$store.dispatch('logOut');
-      this.$router.push({ name: 'login' });
-    },
-  },
-};
