--- front/src/components/SidebarComponent/Sidebar.js
@@ -1,5 +1,0 @@
-export default {
-  name: 'SidebarComponent',
-
-  props: ['activeMenu'],
-};
