+++ front/src/locales/index.js
@@ -1,9 +1,4 @@
-// Import
-import Vue from 'vue';
-import VueI18n from 'vue-i18n';
-
-// Config
-Vue.use(VueI18n);
+import { createI18n } from 'vue-i18n';
 
 // Load messages
 var messages = {
@@ -18,7 +13,7 @@ var messages = {
 };
 
 // Export VueI18n
-export default new VueI18n({
+export default createI18n({
   locale: process.env.VUE_APP_I18N_LOCALE || 'en',
   fallbackLocale: process.env.VUE_APP_I18N_FALLBACK_LOCALE || 'en',
   messages: messages,
