+++ front/src/main.js
@@ -1,30 +1,27 @@
-// Import
-import Vue from 'vue';
+import { createApp } from 'vue';
 import App from './App.vue';
 import router from './router';
 import i18n from './locales';
-import VueMeta from 'vue-meta';
-import store from './store';
-import Notifications from 'vue-notification';
+import { createPinia } from 'pinia';
+import notification from '@kyvg/vue3-notification';
+import authLayout from '@/layouts/auth';
+import adminLayout from '@/layouts/admin';
+
+const pinia = createPinia();
+
+const app = createApp(App);
+app.use(pinia);
+app.use(i18n);
+app.use(notification);
+app.use(router);
 
 // Register layouts
-const authLayout = () => import('./layouts/auth');
-const adminLayout = () => import('./layouts/admin');
-Vue.component('authLayout', authLayout);
-Vue.component('adminLayout', adminLayout);
+app.component('authLayout', authLayout);
+app.component('adminLayout', adminLayout);
 
 // Config
-Vue.config.productionTip = false;
-Vue.prototype.$backendUrl = process.env.VUE_APP_BACKEND_URL;
-Vue.prototype.$appName = process.env.VUE_APP_NAME;
-Vue.use(Notifications);
-Vue.use(VueMeta);
+app.config.globalProperties.config = { productionTip: false };
+app.config.globalProperties.$backendUrl = process.env.VUE_APP_BACKEND_URL;
+app.config.globalProperties.$appName = process.env.VUE_APP_NAME;
 
-// Start Vue Js Instance
-new Vue({
-  el: '#app',
-  router,
-  i18n,
-  store,
-  render: (h) => h(App),
-}).$mount('#app');
+app.mount('#app');
