--- front/src/modules/DashboardModule/Dashboard.js
@@ -1,9 +1,0 @@
-export default {
-  name: 'DashboardModule',
-
-  metaInfo: function () {
-    return {
-      title: this.$t('navbar.dashboard'),
-    };
-  },
-};
