--- front/src/modules/Error404Module/Error404.js
@@ -1,9 +1,0 @@
-export default {
-  name: 'Error404Module',
-
-  metaInfo: function () {
-    return {
-      title: this.$t('notFound.title'),
-    };
-  },
-};
