+++ front/src/modules/Error404Module/index.vue
@@ -1,3 +1,13 @@
 <template src="./Error404.html"></template>
-<script src="./Error404.js"></script>
+<script>
+export default {
+  name: 'Error404Module',
+
+  metaInfo: function () {
+    return {
+      title: this.$t('notFound.title'),
+    };
+  },
+};
+</script>
 <style src="./Error404.css" scoped></style>
