+++ front/src/store/index.js
@@ -1,20 +1,32 @@
-import Vue from 'vue';
-import Vuex from 'vuex';
+import { defineStore } from 'pinia';
 
-Vue.use(Vuex);
-
-import actions from './actions';
 import getters from './getters';
-import mutations from './mutations';
 
-const state = {
-  displayLoader: false,
-  loggedUser: localStorage.getItem('loggedUser') || null,
+const getState = () => {
+  return {
+    displayLoader: false,
+    loggedUser: localStorage.getItem('loggedUser') || null,
+  };
 };
 
-export default new Vuex.Store({
-  state,
-  actions,
-  getters,
-  mutations,
+export const useStore = defineStore('main', {
+  state: getState,
+  getters: getters,
+  actions: {
+    setDisplayLoader(displayLoader) {
+      this.displayLoader = displayLoader;
+    },
+    setLoggedUser(user) {
+      let now = new Date();
+      let expiryDate = new Date();
+      user.expiryDate = expiryDate.setTime(now.getTime() + user.expires_in * 1000);
+
+      localStorage.setItem('loggedUser', JSON.stringify(user));
+      this.loggedUser = JSON.stringify(user);
+    },
+    logOut() {
+      this.loggedUser = null;
+      localStorage.removeItem('loggedUser');
+    },
+  },
 });
